Fixes #2377
In inject's ResourceConfig, renamed objMeta to templateMeta since
it really points to the pod template metadata. And created a new field
objMeta that really points to the main workload (e.g. Deployment) metadata.

Refactored uninject to clean up the labels at both objMeta and
templateMeta. Also it will remove all the labels and annotations that
start with "linkerd.io" except for the "linkerd.io/inject" annotation.

Signed-off-by: Alejandro Pedraza <alejandro@buoyant.io>