Update tap integration test to account for always-on mTLS
Out of all the integration tests (egress, get, stat, tap and
install_test) only in stat and tap do meshed (proxy-to-proxy) connections take
place, which we can test are 100% TLS.

For stat, #2537 already added such check for connections with the
Prometheus pod (connections to other pods are not meshed, apparently).

This PR adds such check for tap.

Fixes #2519

Signed-off-by: Alejandro Pedraza <alejandro@buoyant.io>