Fixes #2720 and 2711 (made more sense to tackle both at once)
This changes the default behavior of `linkerd inject` to not inject the
proxy but just the `linkerd.io/inject: enabled` annotation for the
auto-injector to pick it up (regardless of any namespace annotation).

A new `--advanced` mode was added, which behaves as before, injecting
the proxy in the command output.

The unit tests are running with `--advanced` to avoid any changes in the
fixtures.

I still need to add a new test for non-advanced mode and fix the
integration tests.

Signed-off-by: Alejandro Pedraza <alejandro@buoyant.io>