Create the CA secret during the 'config' stage
This ensures that CA secret is in-sync with the MWC and VWC trust anchors.
Since the MWC and VWC are cluster-scoped, they must be created in the 'config'
stage. Moving the CA secret to the 'config' stage is simpler than trying
to sync it with the MWC and VWC trust anchors.

Signed-off-by: Ivan Sim <ivan@buoyant.io>