New `linkerd upgrade` flag for setting mutating webhook failure policy
Fixes #2852

Added the `--failure-policy` flag for `linkerd upgrade`, that allows
setting the mutating webhook failure policy.

There needs to be at least one proxy injector pod running when the
failure policy is changed to 'Fail': if there's none and the policy is
changed, when the proxy injector tries to start it won't be allowed
because the webhook config will force it to go through an already
existing proxy injector and fail if there's none.

So the policy can be changed to 'Fail` only when:
- Linkerd was already installed with HA
- The CLI version matches the control plane version (so there are no
changes to the proxy-injector that would cause it to be redeployed)
- The control plane is healthy

Note that these checks are performed only the first time that flag is
set.

Another change: when doing `linkerd install --ha` the failure policy for
the validating webhook config is set to Fail.

Still to do: tests.

Signed-off-by: Alejandro Pedraza <alejandro@buoyant.io>