Have `linkerd endpoints` use `Destination.Get`
Fixes #2885

(By popular demand, this supersedes #2937 and refactors `linkerd
endpoints` without creating a separate client for the Destination gRPC
methods)

To recapitulate #2885: we're refactoring `linkerd endpoints` so it hits
directly the `Destination.Get` endpoint, instead of relying on the
Discovery service.

For that, I've created a new `client.go` for Destination and added it to
the `APIClient` interface.

I've also added a `destinationClient` struct that mimics `tapClient`,
and whose common logic has been moved into `stream_client.go`.

Analogously, I added a `destinationServer` struct that mimics
`tapServer`.

Still to do:
- Remove everything related to the Discovery service?
- Fix tests

Signed-off-by: Alejandro Pedraza <alejandro@buoyant.io>